RavenDB-26287: Migrate remaining Npgsql type-loading queries to AST matching + cleanup
